Add files via upload
This demonstrates my dedication to placing my best foot forward to present myself and possibly pitch myself to upcoming companies that are optimal for my candidacy as a new graduate. I've taken the initiative in class to complete the packet tracer course and ISC2 course to obtain my certificate and prepare to sit for official certification. Also, I have final network reports for a local company and residential heat mapping to show the IOT and connections of the residence.